Factor the unconnected HsmManager into an abstract class
This promotes code reuse by other tests: the AbstractHsmManager allows tests
to override only the HsmManager APIs they require.
